Minor extensions.

+++ src/it/cnr/isti/vir/features/IIntValues.java
@@ -2,5 +2,5 @@ package it.cnr.isti.vir.features;
 
 public interface IIntValues extends IArrayValues {
 
-	public float[] getValues();
+	public int[] getValues();
 }
